refactor: cleanup unused helper file
